pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "pycryptodome>=3.23.0",
    "werkzeug>=3.1.3",
]
